--- a/configure.py
+++ b/configure.py
@@ -1992,7 +1992,7 @@ def configure_seastar(build_dir, mode, mode_config):
         '-DSeastar_DEPRECATED_OSTREAM_FORMATTERS=OFF',
         '-DSeastar_UNUSED_RESULT_ERROR=ON',
         '-DCMAKE_EXPORT_COMPILE_COMMANDS=ON',
-        '-DSeastar_SCHEDULING_GROUPS_COUNT=19',
+        '-DSeastar_SCHEDULING_GROUPS_COUNT=20',
         '-DSeastar_IO_URING=ON',
     ]