--- a/gms/feature_service.hh
+++ b/gms/feature_service.hh
@@ -170,6 +170,7 @@ public:
     gms::feature lwt_with_tablets { *this, "LWT_WITH_TABLETS"sv };
     gms::feature repair_msg_split { *this, "REPAIR_MSG_SPLIT"sv };
     gms::feature view_building_coordinator { *this, "VIEW_BUILDING_COORDINATOR"sv };
+    gms::feature driver_service_level { *this, "DRIVER_SERVICE_LEVEL"sv };
 public:
 
     const std::unordered_map<sstring, std::reference_wrapper<feature>>& registered_features() const;