--- a/service/qos/service_level_controller.hh
+++ b/service/qos/service_level_controller.hh
@@ -104,6 +104,7 @@ using update_both_cache_levels = bool_class<class update_both_cache_levels_tag>;
 class service_level_controller : public peering_sharded_service<service_level_controller>, public service::endpoint_lifecycle_subscriber {
 public:
     static inline const int32_t default_shares = 1000;
+    static constexpr auto driver_service_level_name = "driver";
 
     class service_level_distributed_data_accessor {
     public: