--- a/service/storage_service.cc
+++ b/service/storage_service.cc
@@ -7530,6 +7530,11 @@ void storage_service::init_messaging_service() {
                 std::move(muts.begin(), muts.end(), std::back_inserter(mutations));
             }
 
+            auto sl_driver_created_mut = co_await ss._sys_ks.local().get_service_level_driver_created_mutation();
+            if (sl_driver_created_mut) {
+                mutations.push_back(canonical_mutation(*sl_driver_created_mut));
+            }
+
             auto sl_version_mut = co_await ss._sys_ks.local().get_service_levels_version_mutation();
             if (sl_version_mut) {
                 mutations.push_back(canonical_mutation(*sl_version_mut));