--- a/service/topology_coordinator.cc
+++ b/service/topology_coordinator.cc
@@ -3187,6 +3187,10 @@ public:
 };
 
 future<std::optional<group0_guard>> topology_coordinator::maybe_migrate_system_tables(group0_guard guard) {
+    // Some of the upgrades are guarded by _feature_service. If the feature gets enabled,
+    // it's in `topology_coordinator::enable_features` ,so  topology_coordinator will re-run its loop
+    // and `maybe_migrate_system_tables` will be called.
+
     // Check if we can upgrade the view_build_status table to v2, being managed by group0.
     // First we upgrade to an intermediate version v1_5 where we write to both tables, then
     // we upgrade to v2.
@@ -3211,6 +3215,13 @@ future<std::optional<group0_guard>> topology_coordinator::maybe_migrate_system_t
         auto tmptr = get_token_metadata_ptr();
         co_await db::view::view_builder::migrate_to_v2(tmptr, _sys_ks, _sys_ks.query_processor(), _group0.client(), _as, std::move(guard));
         co_return std::nullopt;
+    }
+
+    if (_sl_controller.is_v2() && _feature_service.driver_service_level) {
+        const auto sl_driver_created = co_await _sys_ks.get_service_level_driver_created();
+        if (!sl_driver_created.value_or(false)) {
+            co_return co_await _sl_controller.migrate_to_driver_service_level(std::move(guard), _sys_ks);
+        }
     }
 
     co_return std::move(guard);