--- a/service/topology_coordinator.hh
+++ b/service/topology_coordinator.hh
@@ -19,6 +19,7 @@
 #include "service/endpoint_lifecycle_subscriber.hh"
 #include "service/topology_state_machine.hh"
 #include "db/view/view_building_state.hh"
+#include "service/qos/service_level_controller.hh"
 
 namespace db {
 class system_keyspace;
@@ -90,6 +91,7 @@ future<> run_topology_coordinator(
         std::chrono::milliseconds ring_delay,
         endpoint_lifecycle_notifier& lifecycle_notifier,
         gms::feature_service& feature_service,
+        qos::service_level_controller& sl_controller,
         topology_coordinator_cmd_rpc_tracker& topology_cmd_rpc_tracker);
 
 }