@avikivity 2025-08-21
I'd like the cover letter to contain a description of a manual test that was used to verify not just the inner workings, but that the purpose of the patch was achieved.

1. start a workload (could be a single node single shard cluster)
2. unleash a connection storm, show bad behavior
3. upgrade to new version
4. redo, show good behavior

You can skew the setup to make the bad behavior pronounced, for example create 1000 keyspaces with a table each. I know some drivers will generate O(nr_keyspaces) queries. Have a thousand clients for the connection storm.