@andrzej-jackowski-scylladb 2025-09-01
v6:
 - Updated `process_request_one` to check the current scheduling group, to avoid race condition between setting `service::client_state::service_level_state::USER` and running requests in parallel, just after the service level was changed.