@andrzej-jackowski-scylladb 2025-09-01
v7:
 - Rebased to remove conflicts