@andrzej-jackowski-scylladb 2025-09-01
v8:
 - Cosmetic changes to resolve review comments