@andrzej-jackowski-scylladb 2025-09-02
v9:
 - Added `sl:driver` creation in `raft_initialize_discovery_leader` to increase the feature reliability (as proposed by @nuivall to minimize cases when for some reason `sl:driver is not created) and also make eliminate tests flakiness.

I also analyzed the recent test failures:

- [test_one_big_mutation_corrupted_on_startup.dev.3](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/19428/testReport/junit/cluster.dtest.commitlog_test/TestCommitLog/dev___test_py___dev___test_one_big_mutation_corrupted_on_startup_dev_3) [🔍](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_one_big_mutation_corrupted_on_startup.dev.3) is https://github.com/scylladb/scylladb/issues/25627
- [test_mv_retried_writes_reach_all_replicas.debug.1](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/19428/testReport/junit/cluster.mv/test_mv_admission_control/debug___test_py___debug___test_mv_retried_writes_reach_all_replicas_debug_1) [🔍](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_mv_retried_writes_reach_all_replicas.debug.1) and [test_view_building_with_tablet_move.debug.3](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/19428/testReport/junit/cluster.mv/test_mv_building/debug___test_py___debug___test_view_building_with_tablet_move_debug_3) [🔍](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_view_building_with_tablet_move.debug.3) is https://github.com/scylladb/scylladb/issues/25702
 - `sla_per_user_test.*` failures are expected (just need to adjust the check for the existence of `sl:driver`, I will update dtests when the PR will be ready.
 - All other test failures I expect will be fixed by the `sl:driver` creation in `raft_initialize_discovery_leader`