@andrzej-jackowski-scylladb 2025-09-03
v10:
 - `test_describe.py` - prevent from removing `sl:driver` in test cleanup, it broke other cqlpy tests
- Comment update in `test_service_levels.py`

> #### 🔴 CI State: FAILURE
I reviewed the recent test failures:
 - `cqlpy.test_service_levels.test_scheduling_groups_limit` failed because the node state was invalid due to `test_describe.py` cleanup (fixed in v10)
 - `sla_per_user_test.TestSLA*`  were expected to fail (dtests need adjustment for the new service level)
 - `enterprise_upgrade_test.TestEnterpriseUpgradeFromOSS.test_workload_prioritization_after_upgrade` dtest needs similar fix as `sla_per_user_test`