@andrzej-jackowski-scylladb 2025-09-10
v14:
 - Added special handling of `sl:driver` for `DESC ...`, and `test_desc_driver_service_level`
 - Improved `transport/server.cc` logic to get rid of `co_return co_await ` and increase code clearness

v15:
 - Empty commit, just fixed commit author problem introduced with git rebase