@andrzej-jackowski-scylladb 2025-09-10
v16:
 - Reworked the service level changing logic in `transport/server.cc` because the v14 implementation added ~200 instructions on critical path (it was just a single if, but accessing `static thread_local` variable via `current_scheduling_group()` and comparing sg names).
 - Fixes for `cqlpy.test_service_levels.test_scheduling_groups_limit` and `cluster.auth_cluster.test_raft_service_levels.test_service_levels_upgrade`


**Micro benchmarks of transport/server**
As stated above, v14 implementation introduced an overhead of ~200 instructions on handling `QUERY` requests, measured by the [perf_cql_raw](https://github.com/scylladb/scylladb/commit/67fbe35833e2d23a8e9c2dcb5e04580231d8ec96) benchmark.

Baseline (6116f9e11b4c00de60c2e8b11b93edb66c39a3b2):
```
instructions_per_op:
	mean=   108075.64 standard-deviation=633.52
	median= 108011.68 median-absolute-deviation=64.79
	maximum=114346.67 minimum=107996.44
```

v14 (11849360966460067a4282388f24db724b5b7f1b):
```
instructions_per_op:
	mean=   108235.52 standard-deviation=662.70
	median= 108166.23 median-absolute-deviation=70.11
	maximum=114795.26 minimum=108144.00
```

Final version (e3b9ec070f5cd3536d411665d53c9de689c6ae0a):
```
instructions_per_op:
	mean=   108039.50 standard-deviation=633.56
	median= 107970.67 median-absolute-deviation=70.02
	maximum=114309.34 minimum=107950.68
```

The experiments above were repeated many times, and the results were always similar. The difference was also visible after refreshing pgo profiles.