@andrzej-jackowski-scylladb 2025-09-10
v17:
 - Only cosmetic code / comments changes