@andrzej-jackowski-scylladb 2025-09-10
@piodul, @Jadw1, please review the recent version. The following major changes were introduced since your last review:
- Creation of `sl:driver` was moved to `topology_coordinator` and additionally it is created in `raft_initialize_discovery_leader` for new systems and tests.
- Special handling of `DESC ...` was added to make sure `sl:driver` works correctly in backups.
- `transport/server.cc` logic was reworked to reduce complexity and avoid overheads
- Additional tests added to cover the new changes.
- Some (flaky) tests were fixed.