@andrzej-jackowski-scylladb 2025-09-17
> [test_view_building_with_tablet_move.debug.1](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/20011/testReport/junit/cluster.mv/test_mv_building/debug___test_py___debug___test_view_building_with_tablet_move_debug_1) [🔍](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_view_building_with_tablet_move.debug.1)

https://github.com/scylladb/scylladb/issues/25804