@andrzej-jackowski-scylladb 2025-09-17
v19:
 - Only commit messages / comments fixes